% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.ecolRxC}
\alias{print.summary.ecolRxC}
\title{Print a summary of a summary.ecolRxC object}
\usage{
\method{print}{summary.ecolRxC}(x, ..., margins = TRUE, digits = 2)
}
\arguments{
\item{x}{An \code{summary.ecolRxC} class object.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{margins}{A TRUE/FALSE argument informing if the margins of the transition matrix should be displayed. Default TRUE.}

\item{digits}{Integer indicating the number of decimal places to be shown. Default, 2.}
}
\value{
{No return value, called for side effects.}
}
\description{
Print method for \code{summary.ecolRxC} objects
}
